🔥Remove and rename runinfo modal
